+++ src/loaders/core-modules.ts
@@ -1,0 +1,13 @@
+import '../cards/SummaryCard';
+import '../cards/LightsGroupCard';
+import '../cards/CoversGroupCard';
+import '../cards/BatteriesCard';
+import '../cards/AreaNavigationCard';
+import '../cards/CameraCard';
+import '../cards/WebrtcCameraCard';
+import '../views/OverviewViewStrategy';
+
+export { Registry } from '../Registry';
+export { getVisibleAreasFromHass } from '../utils/name-utils';
+export { localize } from '../utils/localize';
+export { withUnavailableEntitiesHidden } from '../utils/availability-utils';