+++ src/loaders/view-modules.ts
@@ -1,0 +1,6 @@
+import '../views/LightsViewStrategy';
+import '../views/CoversViewStrategy';
+import '../views/SecurityViewStrategy';
+import '../views/BatteriesViewStrategy';
+import '../views/ClimateViewStrategy';
+import '../views/RoomViewStrategy';